Refactor workspace access control
